treewide: Build everything into the kernel image
It's really a big pain to have modules when making standalone kernel
images as modules usually mismatch with either signature or actual
contents from compile to compile variance. Drop all modules and bake
everything into the kernel. This saves us a lot of time and effort
when testing kernel builds as now I don't have to flash an entire
build just to test the kernel.

Signed-off-by: Sourajit Karmakar <sourajit.karmakar@gmail.com>